+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Fix bug where wrapInit wasn't called on v2 callable functions. (#1634)
+- Add support for callable function to return streaming response (#1629)